[CDG-73][CDG-74] - Layout adjustments
