+++ src/components/RelationsSelect/RelationSelect.style.ts
@@ -1,0 +1,10 @@
+import { styled } from "@stitches/react";
+
+export const SelectContainer = styled("div", {
+  display: "flex",
+  flexDirection: "column",
+  maxWidth: "160px",
+  justifyContent:"center",
+  gap: "4px",
+});
+
